address review concerns
