+++ src/components/views/Experiment.js
@@ -127,7 +127,7 @@ export default class extends React.Component {
             <div>
                 <article id="experiment">
                     <h2>{this.props.name || this.props.slug}</h2>
-                    <div className="experiment-info">
+                    <div>
                         <section id="experiment-details">
                             <h3>Details</h3>
                             {maybeDescription}
