+++ src/components/views/styl/Experiment.styl
@@ -123,8 +123,3 @@
         margin-bottom: 0;
     }
 }
-
-.experiment-info {
-    border-bottom: 1px solid $gray;
-    padding-bottom: 15px;
-}
