+++ src/components/views/styl/ExperimentsTable.styl
@@ -8,19 +8,9 @@
     a, a:visited {
         color: #00f;
         text-decoration: none;
-        padding: 0 10px;
     }
 
     .experiment-creation-date {
         width: 20%;
-    }
-
-    tr {
-        padding: 10px 0 9px;
-    }
-
-    tbody tr:nth-child(odd) {
-        background: linear-gradient(to right, rgba(150, 150, 150, 0.15), rgb(255, 255, 255));
-        border-bottom: 0.5px solid rgba(0, 0, 0, 0.25);
     }
 }
